requirements.txt
flask
flask-cors
google-generativeai
pymupdf
pillow
